+++ public/js/models/settings.js
@@ -1,0 +1,8 @@
+var Backbone = require('backbone');
+
+var Settings = Backbone.Model.extend({
+    url: '/_/settings'
+});
+
+
+module.exports = new Settings();